Apparently pull request templates don't work in a user-friendly way on GH
You have to specify `?template=<template-file.md>` on the URL as you
create the PR to get the template you want; there is not a selection of
choices presented to the user instead, as there are with multiple issue
templates.

We return to the single PR template for now.
I stacked in both new submission and compute expansion checklists, with
instructions for the user as a comment.